allow client to pass session_id for multiple pushers by 1 service worker
A service worker can only have 1 push subscription, but for web clients
that handle multiple matrix accounts like hydrogen, we want to be able
to enable push for every account. We can do that by setting a pusher
with the same web push details for each account, but we can't really
tell which account a push notification is for, so forward a session_id
set on the pusher data.